requirements.txt
# Brogue MVP Dependencies
# See: docs/architecture/MVP_DEVELOPMENT_GUIDELINES.md

# ===== Core Infrastructure =====
python>=3.10

# ===== UI =====
textual>=0.45.0            # Terminal UI framework
rich>=13.7.0               # Rich terminal output

# ===== Data & Content =====
pyyaml>=6.0                # YAML content loading (recipes, monsters)
pydantic>=2.0              # Type-safe models (optional)

# ===== Development & Testing =====
pytest>=7.4.0
pytest-cov>=4.1.0

# ===== Development Tools =====
textual-dev>=1.0.0         # Textual development tools